Unmerge the models/rules files.
